requirements.txt
playwright
python-telegram-bot[job-queue]>=20.0
apscheduler
python-dotenv
beautifulsoup4
